requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
lxml
openai
pydub
moviepy
google-genai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
numpy
opencv-python-headless
google-generativeai
yt-dlp
pysubs2
openai-whisper
Flask>=2.0
gunicorn>=20.0
flask-cors>=3.0
python-multipart
aiofiles>=23.1.0